Fix tests wrt moving k-opt heuristic functions
