+++ procrustes/test/test_utils.py
@@ -23,7 +23,11 @@
 """Utils module for Procrustes."""
 
 import numpy as np
-from procrustes.utils import _hide_zero_padding, _scale_array, _translate_array, _zero_padding
+from numpy.testing import assert_equal, assert_raises
+
+from procrustes.utils import (_hide_zero_padding, _scale_array,
+                              _translate_array, _zero_padding, error,
+                              kopt_heuristic_double, kopt_heuristic_single)
 
 
 def test_zero_padding_rows():
@@ -320,3 +324,68 @@ def test_scale_array():
     # array_trans_scale should be identical to array after the above analysis
     expected = array_a
     assert (abs(predicted - expected) < 1.e-10).all()
+
+
+def test_kopt_heuristic_single():
+    r"""Test k-opt heuristic search algorithm."""
+    arr_a = np.array([[3, 6, 1, 0, 7],
+                      [4, 5, 2, 7, 6],
+                      [8, 6, 6, 1, 7],
+                      [4, 4, 7, 9, 4],
+                      [4, 8, 0, 3, 1]])
+    arr_b = np.array([[1, 8, 0, 4, 3],
+                      [6, 5, 2, 4, 7],
+                      [7, 6, 6, 8, 1],
+                      [7, 6, 1, 3, 0],
+                      [4, 4, 7, 4, 9]])
+    perm_guess = np.array([[0, 0, 1, 0, 0],
+                           [1, 0, 0, 0, 0],
+                           [0, 0, 0, 1, 0],
+                           [0, 0, 0, 0, 1],
+                           [0, 1, 0, 0, 0]])
+    perm_exact = np.array([[0, 0, 0, 1, 0],
+                           [0, 1, 0, 0, 0],
+                           [0, 0, 1, 0, 0],
+                           [0, 0, 0, 0, 1],
+                           [1, 0, 0, 0, 0]])
+    error_old = error(arr_a, arr_b, perm_guess, perm_guess)
+    perm, kopt_error = kopt_heuristic_single(perm_guess, arr_a, arr_b,
+                                             error_old, 3, kopt_tol=1.e-8)
+    assert_equal(perm, perm_exact)
+    assert kopt_error == 0
+    # test the error exceptions
+    assert_raises(ValueError, kopt_heuristic_single, perm_guess,
+                  arr_a, arr_b, error_old, 1, kopt_tol=1.e-8)
+
+
+def test_kopt_heuristic_double():
+    r"""Test double sided k-opt heuristic search algorithm."""
+    np.random.seed(998)
+    arr_b = np.random.randint(low=-10, high=10, size=(4, 3)).astype(np.float)
+    perm1 = np.array([[0., 0., 0., 1.],
+                      [0., 1., 0., 0.],
+                      [1., 0., 0., 0.],
+                      [0., 0., 1., 0.]])
+    perm2 = np.array([[0., 0., 1.],
+                      [1., 0., 0.],
+                      [0., 1., 0.]])
+    arr_a = np.linalg.multi_dot([perm1.T, arr_b, perm2])
+    # shuffle the permutation matrices
+    perm1_shuff = np.array([[0., 0., 0., 1.],
+                            [1., 0., 0., 0.],
+                            [0., 1., 0., 0.],
+                            [0., 0., 1., 0.]])
+    perm2_shuff = np.array([[1., 0., 0.],
+                            [0., 0., 1.],
+                            [0., 1., 0.]])
+    e_opt = error(arr_b, arr_a, perm1_shuff.T, perm2_shuff)
+    perm_left, perm_right, kopt_error = kopt_heuristic_double(perm_p=perm1_shuff,
+                                                              perm_q=perm2_shuff,
+                                                              array_m=arr_a, array_n=arr_b,
+                                                              ref_error=e_opt, kopt_k=4,
+                                                              kopt_tol=1.e-8)
+    _, _, kopt_error = kopt_heuristic_double(perm_p=perm_left, perm_q=perm_right,
+                                             array_m=arr_a, array_n=arr_b,
+                                             ref_error=e_opt, kopt_k=3, kopt_tol=1.e-8)
+    assert kopt_error <= e_opt
+    assert kopt_error == 0
